+++ paramdb/__init__.py
@@ -2,6 +2,7 @@
 Database for storing and retrieving QPU parameters during quantum control experiments.
 """
 
+from paramdb._keys import CLASS_NAME_KEY, PARAMLIST_ITEMS_KEY, LAST_UPDATED_KEY
 from paramdb._param_data._param_data import ParamData
 from paramdb._param_data._dataclasses import Param, Struct
 from paramdb._param_data._collections import ParamList, ParamDict
@@ -9,6 +10,9 @@ from paramdb._param_data._type_mixins import ParentType, RootType
 from paramdb._database import ParamDB, CommitEntry
 
 __all__ = [
+    "CLASS_NAME_KEY",
+    "PARAMLIST_ITEMS_KEY",
+    "LAST_UPDATED_KEY",
     "ParamData",
     "Param",
     "Struct",